Fixed many bugs. Separated the example Scenes to a separate .cpp file.

+++ jniLibs/MobileRT/src/main/cpp/MobileRT/Sampler.hpp
@@ -13,7 +13,7 @@ namespace MobileRT {
     public:
         std::atomic<uint64_t> sample_ {0};
         const uint64_t domainSize_ {0};
-        uint64_t samples_ {0};
+        unsigned int samples_ {0};
 
     public:
         explicit Sampler(uint64_t domainSize,
